Fix detection of commits when using a local repository.
When using a local repository (instead of one cloned from a Internet
server), RefactoringMiner fails to detect any commit. This is due to
detection of the refs, which was just considering remote references.
However, when using a repository that was not cloned, but created
locally (using 'git init'), the expected refs is different.

The fix consists of also checking for a local refs ('refs/heads/')
when walking within the repository.

Looking at other references to REMOTE_REFS_PREFIX, the public method
'countCommits(Repository repository, String branch)' probably is also
broken. However, as it is currently not used by RefactoringMiner,
I did not tested or changed it.